+++ seerep_srv/seerep_core_fb/include/seerep_core_fb/core_fb_conversion.h
@@ -7,24 +7,18 @@
 // grpc / flatbuffer
 #include <flatbuffers/flatbuffers.h>
 #include <flatbuffers/grpc.h>
-#include <seerep-msgs/camera_intrinsics_generated.h>
-#include <seerep-msgs/camera_intrinsics_query_generated.h>
+#include <seerep_msgs/camera_intrinsics_generated.h>
+#include <seerep_msgs/camera_intrinsics_query_generated.h>
 // seerep-msgs
-#include <seerep-msgs/camera_intrinsics.h>
-#include <seerep-msgs/camera_intrinsics_query.h>
-#include <seerep-msgs/datatype_generated.h>
-#include <seerep-msgs/image_generated.h>
-#include <seerep-msgs/point_cloud_2_generated.h>
-#include <seerep-msgs/point_stamped_generated.h>
-#include <seerep-msgs/query_generated.h>
-#include <seerep-msgs/query_instance_generated.h>
-#include <seerep-msgs/region_of_interest.h>
+#include <seerep_msgs/camera_intrinsics.h>
+#include <seerep_msgs/camera_intrinsics_query.h>
 #include <seerep_msgs/datatype_generated.h>
 #include <seerep_msgs/image_generated.h>
 #include <seerep_msgs/point_cloud_2_generated.h>
 #include <seerep_msgs/point_stamped_generated.h>
 #include <seerep_msgs/query_generated.h>
 #include <seerep_msgs/query_instance_generated.h>
+#include <seerep_msgs/region_of_interest.h>
 #include <seerep_msgs/transform_stamped_generated.h>
 #include <seerep_msgs/transform_stamped_query_generated.h>
 #include <seerep_msgs/uuids_per_project_generated.h>
